Model knows map information.
Model can provide information from Static DB about static map.
Added new message - MapSnapshotRequestedMessage, as a request from
View for current map snapshot (Model provides it, [from cache]).
Added translation from Model::Map to Model::WorldSnapshot, to separate
Model's structures and Snapshot structures (which are passed to View).
QtRenderer uses autopositioning of scene, so no more width and height
are provided. There are cons for this approach - if track appears even
slightly out of current map - scene moves. It's not common case, but can
appear.
Pros of this solution is that view is centered on the observation area,
does not matter where it is.